修改WebSocket错误提示 解决server接入器退出bug

+++ ClientDemo_cpp_version/typedef.h
@@ -82,6 +82,9 @@ void Exit(int sign)
     inputPipe.write("stop\n",5);
     cout << "Interrupt" << endl;
     sleep(1);
-    kill(pid, SIGABRT);
+    if(sign)
+    {
+        kill(pid, SIGABRT);
+    }
     exit(EXIT_FAILURE);
 }
